+++ packages/p2p/src/components/my-ads/my-ads-delete-modal.scss
