+++ packages/p2p/src/pages/my-ads/edit-ad/edit-ad-form-payment-methods/edit-ad-form-payment-methods.jsx
@@ -3,8 +3,8 @@ import { observer } from 'mobx-react-lite';
 import { useStores } from 'Stores';
 import PaymentMethodCard from 'Components/my-profile/payment-methods/payment-method-card';
 import { localize } from 'Components/i18next';
-import BuyAdPaymentMethodsList from './buy-ad-payment-methods-list/buy-ad-payment-methods-list.jsx';
-import SellAdPaymentMethodsList from './sell-ad-payment-methods-list.jsx';
+import BuyAdPaymentMethodsList from '../../buy-ad-payment-methods-list/buy-ad-payment-methods-list.jsx';
+import SellAdPaymentMethodsList from '../../sell-ad-payment-methods-list.jsx';
 
 const EditAdFormPaymentMethods = ({ is_sell_advert, selected_methods, setSelectedMethods, touched }) => {
     const { my_ads_store, my_profile_store } = useStores();
