+++ packages/p2p/src/pages/my-ads/edit-ad/edit-ad-form.jsx
@@ -12,10 +12,10 @@ import { buy_sell } from 'Constants/buy-sell';
 import { useStores } from 'Stores';
 import { ad_type } from 'Constants/floating-rate';
 import FloatingRate from 'Components/floating-rate';
-import EditAdCancelModal from './edit-ad-cancel-modal.jsx';
+import EditAdCancelModal from 'Components/my-ads/edit-ad-cancel-modal.jsx';
 import { generateErrorDialogTitle, generateErrorDialogBody } from 'Utils/adverts';
-import EditAdFormPaymentMethods from './edit-ad-form-payment-methods.jsx';
-import CreateAdAddPaymentMethodModal from './create-ad-add-payment-method-modal.jsx';
+import EditAdFormPaymentMethods from './edit-ad-form-payment-methods/edit-ad-form-payment-methods.jsx';
+import CreateAdAddPaymentMethodModal from 'Components/my-ads/create-ad-add-payment-method-modal.jsx';
 import EditAdSummary from './edit-ad-summary.jsx';
 
 const EditAdFormWrapper = ({ children }) => {
