+++ packages/p2p/src/pages/my-ads/my-ads-table.jsx
@@ -10,11 +10,11 @@ import { TableError } from 'Components/table/table-error.jsx';
 import { ad_type } from 'Constants/floating-rate';
 import { useStores } from 'Stores';
 import { generateErrorDialogTitle } from 'Utils/adverts';
-import MyAdsDeleteModal from './my-ads-delete-modal.jsx';
-import MyAdsFloatingRateSwitchModal from './my-ads-floating-rate-switch-modal.jsx';
+import MyAdsDeleteModal from 'Components/my-ads/my-ads-delete-modal.jsx';
+import MyAdsFloatingRateSwitchModal from 'Components/my-ads/my-ads-floating-rate-switch-modal.jsx';
 import MyAdsRowRenderer from './my-ads-row-renderer.jsx';
-import QuickAddModal from './quick-add-modal.jsx';
-import AdExceedsDailyLimitModal from './ad-exceeds-daily-limit-modal.jsx';
+import QuickAddModal from 'Components/my-ads/quick-add-modal.jsx';
+import AdExceedsDailyLimitModal from 'Components/my-ads/ad-exceeds-daily-limit-modal.jsx';
 
 const getHeaders = offered_currency => [
     { text: localize('Ad ID') },
