+++ packages/p2p/src/pages/my-ads/my-ads.jsx
@@ -5,8 +5,8 @@ import { observer } from 'mobx-react-lite';
 import { localize } from 'Components/i18next';
 import { useStores } from 'Stores';
 import { TableError } from 'Components/table/table-error.jsx';
-import CreateAd from './create-ad.jsx';
-import EditAd from './edit-ad.jsx';
+import CreateAd from './create-ad/create-ad.jsx';
+import EditAd from './edit-ad/edit-ad.jsx';
 import MyAdsTable from './my-ads-table.jsx';
 import Verification from 'Components/verification/verification.jsx';
 import './my-ads.scss';
